add CSR screenshot example
