--- packages/template/src/lib/Hi.svelte
@@ -1,5 +1,0 @@
-<script lang="ts">
-  export let name: string
-</script>
-
-Hello, {name}!
